Make engine subprocess-based
The bindings I was using require the latest version of VIPS, which
is different from the one we have in production. Those bindings
are particularly annoying to install, too.

This simplifies the engine, making it more like the other ones,
caring only about converting the image to a supported format.

Bug: T119184
Change-Id: I51502cd0bee47691ca0b40cfd6a43eb9d73a4ef5